* Updated size functions in KaRL containers Map and Vector to use size_t instead of unsigned int, so warnings do not occur during 64 bit Visual Studio compilation * Updated to version 1.5.4
git-svn-id: http://madara.googlecode.com/svn/trunk@1217 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/knowledge_engine/Map.h
@@ -56,7 +56,7 @@ namespace Madara
        * Returns the size of the map
        * @return size of the map
        **/
-      unsigned int size (void);
+      size_t size (void);
       
       /**
        * Syncs the keys from the knowledge base. This can be useful
